Add support for Firefox as a RemoteWebDriver.

+++ src/Driver.java
@@ -13,20 +13,26 @@ import org.apache.commons.cli.CommandLine;
 import org.apache.commons.cli.CommandLineParser;
 import org.apache.commons.cli.GnuParser;
 import org.apache.commons.cli.HelpFormatter;
+import org.apache.commons.cli.Option;
 import org.apache.commons.cli.OptionBuilder;
+import org.apache.commons.cli.OptionGroup;
 import org.apache.commons.cli.Options;
 import org.apache.commons.cli.ParseException;
 import org.openqa.selenium.By;
 import org.openqa.selenium.NoSuchElementException;
 import org.openqa.selenium.OutputType;
+import org.openqa.selenium.TakesScreenshot;
 import org.openqa.selenium.WebElement;
+import org.openqa.selenium.firefox.FirefoxDriver;
 import org.openqa.selenium.phantomjs.PhantomJSDriver;
 import org.openqa.selenium.remote.DesiredCapabilities;
+import org.openqa.selenium.remote.RemoteWebDriver;
 import org.openqa.selenium.remote.ScreenshotException;
 
 public class Driver 
 {
-	private final PhantomJSDriver dri;
+	private final boolean useFirefox;
+	private final RemoteWebDriver dri;
 	public static final String programName = "so-chatbot-driver";
 	public static boolean displayScreenshots = false;
 	
@@ -37,9 +43,16 @@ public class Driver
 		Options options = new Options();
 		CommandLine line = null;
 		HelpFormatter hf = new HelpFormatter();
+		OptionGroup og = new OptionGroup();
+		Option opjs = OptionBuilder.withLongOpt("phantomjs").withDescription("Use PhantomJS").create('j');
+		Option off = OptionBuilder.withLongOpt("firefox").withDescription("Use Firefox").create('f');
+		og.setRequired(true);
+		og.addOption(opjs);
+		og.addOption(off);
 		
 		options.addOption("s", "screenshot", false, "Display screenshots at checkpoints");
 		options.addOption("h",  "help", false, "Print this message" );
+		options.addOptionGroup(og);
 		options.addOption(OptionBuilder.withLongOpt( "bot-script" ).withDescription( "Required: Bot script file name" ).hasArg().withArgName("FILE").create('b'));
 		options.addOption("u", "username", true, "Required: StackExchange username or email");
 		options.addOption("p", "password", true, "Required: StackExchange password");
@@ -74,7 +87,17 @@ public class Driver
 			}
 		}
 		
-		Driver d = new Driver();
+		if(!line.hasOption("j") && !line.hasOption("f"))
+		{
+			hf.printHelp(programName, options);
+			return;
+		}
+		
+		boolean hff = false;
+		if(line.hasOption("f"))
+			hff = true;
+		
+		Driver d = new Driver(hff);
 		try
 		{
 			d.go(line.getOptionValue('u'), line.getOptionValue('p'), line.getOptionValue('b'), line.getOptionValue('c'));
@@ -101,10 +124,11 @@ public class Driver
 	}
 	public void go(String s_username, String s_password, String scriptPath, String chatURL) throws InterruptedException, IOException, FileNotFoundException
 	{
+		TakesScreenshot ts = (TakesScreenshot) dri;
 		dri.navigate().to(chatURL);
 		Thread.sleep(rand(5000, 10000));
 		System.out.println(dri.getCurrentUrl() + ": In the chatroom! Determining if we need to login...");
-		ImagePanel.displayImage(dri.getScreenshotAs(OutputType.BASE64));
+		ImagePanel.displayImage(ts.getScreenshotAs(OutputType.BASE64));
 		
 		WebElement loginLink = null;
 		try
@@ -123,7 +147,7 @@ public class Driver
 			//Need to login
 			Thread.sleep(rand(5000, 6500));
 			System.out.println(dri.getCurrentUrl() + ": Performing stage 1 chat-login auth link");
-			ImagePanel.displayImage(dri.getScreenshotAs(OutputType.BASE64));
+			ImagePanel.displayImage(ts.getScreenshotAs(OutputType.BASE64));
 			WebElement midpointLink = null;
 			try
 			{
@@ -139,11 +163,11 @@ public class Driver
 			{
 				Thread.sleep(rand(3000, 6500));
 				System.out.println(dri.getCurrentUrl() + ": We should be at the page that asks you to pick a sign-in method now...");
-				ImagePanel.displayImage(dri.getScreenshotAs(OutputType.BASE64));
+				ImagePanel.displayImage(ts.getScreenshotAs(OutputType.BASE64));
 				
 				dri.executeScript("openid.signin('stack_exchange');");
 				Thread.sleep(rand(2000, 6500));
-				ImagePanel.displayImage(dri.getScreenshotAs(OutputType.BASE64));
+				ImagePanel.displayImage(ts.getScreenshotAs(OutputType.BASE64));
 				System.out.println(dri.getCurrentUrl() + ": We should be at the sign-in page now...");
 				
 				dri.switchTo().frame(dri.findElementById("affiliate-signin-iframe"));
@@ -153,20 +177,20 @@ public class Driver
 				WebElement password = dri.findElement(By.id("password"));
 				password.click();
 				password.sendKeys(s_password);
-				ImagePanel.displayImage(dri.getScreenshotAs(OutputType.BASE64));
+				ImagePanel.displayImage(ts.getScreenshotAs(OutputType.BASE64));
 				WebElement submitButton = dri.findElement(By.className("affiliate-button"));
 				if(submitButton == null)
 					throw new IOException("Couldn't find the affiliate-button classed submit button!");
 				submitButton.click();
 				Thread.sleep(rand(5000, 6500));
 				System.out.println(dri.getCurrentUrl() + ": We should be logged in now...");
-				ImagePanel.displayImage(dri.getScreenshotAs(OutputType.BASE64));
+				ImagePanel.displayImage(ts.getScreenshotAs(OutputType.BASE64));
 				dri.switchTo().defaultContent();
 				
 				dri.navigate().to(chatURL);
 				Thread.sleep(rand(8000, 10000));
 				System.out.println(dri.getCurrentUrl() + ": In the chatroom! Authenticating...");
-				ImagePanel.displayImage(dri.getScreenshotAs(OutputType.BASE64));
+				ImagePanel.displayImage(ts.getScreenshotAs(OutputType.BASE64));
 				
 				loginLink = dri.findElement(By.xpath("//a[starts-with(@href, '/login/global') and text() = 'logged in' and not(ancestor::div[contains(@style,'display:none')]) and not(ancestor::div[contains(@style,'display: none')])]"));
 				loginLink.click();
@@ -181,7 +205,7 @@ public class Driver
 		dri.executeScript(content);
 		
 		System.out.println("Running...");
-		ImagePanel.displayImage(dri.getScreenshotAs(OutputType.BASE64));
+		ImagePanel.displayImage(ts.getScreenshotAs(OutputType.BASE64));
 		BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
 		for(String input = br.readLine(); input != null; input = br.readLine())
 		{
@@ -197,15 +221,25 @@ public class Driver
 	  return StandardCharsets.UTF_8.decode(ByteBuffer.wrap(encoded)).toString();
 	}
 	
-	public Driver() 
+	public Driver(boolean hff) 
 	{
-		File fil = new File("cookies.txt");
-		fil.delete();
-		DesiredCapabilities caps = new DesiredCapabilities(DesiredCapabilities.phantomjs());
-		//caps.setCapability("phantomjs.page.settings.userAgent", "Mozilla/5.0 (Windows NT 6.2; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/28.0.1500.95 Safari/537.36");
-		caps.setCapability("phantomjs.page.settings.userAgent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_6_8) AppleWebKit/537.13+ (KHTML, like Gecko) Version/5.1.7 Safari/534.57.2");
-		caps.setCapability("phantomjs.cli.args", new String[] { "--cookies-file=" + fil.getAbsolutePath() });
-		dri = new PhantomJSDriver(caps);
+		this.useFirefox = hff;
+		if(useFirefox)
+		{
+			//Xvfb needs to be started if this is headless Linux
+			dri = new FirefoxDriver();
+		}
+		else
+		{
+			File fil = new File("cookies.txt");
+			fil.delete();
+			DesiredCapabilities caps = new DesiredCapabilities(DesiredCapabilities.phantomjs());
+			//caps.setCapability("phantomjs.page.settings.userAgent", "Mozilla/5.0 (Windows NT 6.2; WOW64) AppleWebKit/537.36 (KHTML, like Gecko) Chrome/28.0.1500.95 Safari/537.36");
+			caps.setCapability("phantomjs.page.settings.userAgent", "Mozilla/5.0 (Macintosh; Intel Mac OS X 10_6_8) AppleWebKit/537.13+ (KHTML, like Gecko) Version/5.1.7 Safari/534.57.2");
+			caps.setCapability("phantomjs.cli.args", new String[] { "--cookies-file=" + fil.getAbsolutePath() });
+			dri = new PhantomJSDriver(caps);
+		}
+
 		dri.manage().deleteAllCookies();
 		dri.manage().timeouts().implicitlyWait(10, TimeUnit.SECONDS);
 	}
